pyproject.toml
[project]
dependencies = [
  'astropy == 5.2.2',
  'astropy-healpix == 0.7',
  'astroplan == 0.8',
  'joblib == 1.2.0',
  'healpy == 1.16.2',
  'mocpy == 0.12.3',
  'ligo.skymap == 1.0.7',
  'numba == 0.57.0',
  'numba-progress == 0.0.4',
  'numpy == 1.24.3',
  'tqdm == 4.65.0',
]